# server/openapi/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "Get user if exists, create new user if not",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "201": {
            "description": "New user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        },
        "tags": ["users"]
      },
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get a specific user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        },
        "tags": ["users"]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        },
        "tags": ["users"]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete user",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["users"]
      }
    },
    "/users/toggleWatchList": {
      "patch": {
        "operationId": "UsersController_toggleWatchList",
        "summary": "Add or remove item from watch list",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ToggleWatchListDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns watch list array containing 0-n ticker symbols"
          }
        },
        "tags": ["users"]
      }
    },
    "/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "summary": "Create a new transaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTransactionDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "number" } }
            }
          },
          "400": {
            "description": "Not enough shares or money to make the transaction"
          },
          "500": { "description": "Internal Server Error" }
        },
        "tags": ["transactions"]
      },
      "get": {
        "operationId": "TransactionsController_findAll",
        "summary": "Get all user transactions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          }
        },
        "tags": ["transactions"]
      }
    },
    "/stock/candle/{ticker}": {
      "get": {
        "operationId": "StockController_candle",
        "summary": "Get candle data for a specific ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "resolution",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Candle" }
                }
              }
            }
          }
        },
        "tags": ["stock"]
      }
    },
    "/stock/news/{ticker}": {
      "get": {
        "operationId": "StockController_news",
        "summary": "Get related news articles",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Article" }
                }
              }
            }
          }
        },
        "tags": ["stock"]
      }
    },
    "/stock/quote/{ticker}": {
      "get": {
        "operationId": "StockController_quote",
        "summary": "Get the latest price",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Quote" }
              }
            }
          }
        },
        "tags": ["stock"]
      }
    },
    "/stock/search": {
      "get": {
        "operationId": "StockController_search",
        "summary": "Search for companies",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Ticker" }
                }
              }
            }
          }
        },
        "tags": ["stock"]
      }
    }
  },
  "info": {
    "title": "MarketSim",
    "description": "The marketsim api",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "nickname": { "type": "string" },
          "picture": { "type": "string" },
          "email": { "type": "string" },
          "phone_number": { "type": "string" },
          "address": { "type": "string" },
          "sub": { "type": "string" }
        },
        "required": ["email", "sub"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "balance": { "type": "number", "default": 1000000 },
          "telephone": { "type": "string" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "nickname": { "type": "string" },
          "address": { "type": "string" },
          "picture": { "type": "string" },
          "watch_list": {
            "default": [],
            "type": "array",
            "items": { "type": "string" }
          },
          "created_on": { "format": "date-time", "type": "string" },
          "updated_on": { "format": "date-time", "type": "string" }
        },
        "required": ["id", "balance", "watch_list", "created_on", "updated_on"]
      },
      "ToggleWatchListDto": {
        "type": "object",
        "properties": {
          "ticker": { "type": "string" },
          "isWatched": { "type": "boolean" }
        },
        "required": ["ticker", "isWatched"]
      },
      "UpdateUserDto": { "type": "object", "properties": {} },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "quantity": { "type": "number" },
          "symbol": { "type": "string" },
          "type": { "type": "string" }
        },
        "required": ["quantity", "symbol", "type"]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "symbol": { "type": "string" },
          "quantity": { "type": "number" },
          "price": { "type": "number" },
          "type": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" }
        },
        "required": ["symbol", "quantity", "price", "type"]
      },
      "Candle": {
        "type": "object",
        "properties": {
          "c": {
            "description": "List of close prices",
            "type": "array",
            "items": { "type": "number" }
          },
          "h": {
            "description": "List of high prices",
            "type": "array",
            "items": { "type": "number" }
          },
          "l": {
            "description": "List of low prices",
            "type": "array",
            "items": { "type": "number" }
          },
          "o": {
            "description": "List of open prices",
            "type": "array",
            "items": { "type": "number" }
          },
          "s": {
            "type": "string",
            "description": "Response status: ok or no_data"
          },
          "t": {
            "description": "List of timestamps",
            "type": "array",
            "items": { "type": "number" }
          },
          "v": {
            "description": "List of volume data",
            "type": "array",
            "items": { "type": "number" }
          }
        },
        "required": ["c", "h", "l", "o", "s", "t", "v"]
      },
      "Article": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "publisher": {
            "type": "object",
            "properties": {
              "name": { "required": true, "type": "string" },
              "homepage_url": { "required": true, "type": "string" },
              "logo_url": { "required": true, "type": "string" },
              "favicon_url": { "required": true, "type": "string" }
            }
          },
          "title": { "type": "string" },
          "author": { "type": "string" },
          "article_url": { "type": "string" },
          "tickers": { "type": "array", "items": { "type": "string" } },
          "image_url": { "type": "string" },
          "description": { "type": "string" },
          "keywords": { "type": "array", "items": { "type": "string" } }
        },
        "required": [
          "id",
          "publisher",
          "title",
          "author",
          "article_url",
          "tickers",
          "image_url",
          "description",
          "keywords"
        ]
      },
      "Quote": {
        "type": "object",
        "properties": {
          "c": { "type": "number", "description": "Current price" },
          "d": { "type": "number", "description": "Change" },
          "dp": { "type": "string", "description": "Percent change" },
          "h": { "type": "number", "description": "High price of the day" },
          "l": { "type": "number", "description": "Low price of the day" },
          "o": { "type": "number", "description": "Price at open" },
          "pc": { "type": "number", "description": "Previous close price" },
          "t": { "type": "number", "description": "Unix timestamp" }
        },
        "required": ["c", "d", "dp", "h", "l", "o", "pc", "t"]
      },
      "Ticker": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "symbol": { "type": "string" },
          "description": { "type": "string" }
        }
      }
    }
  }
}